Added check before moving file

+++ install.sh
@@ -1,3 +1,5 @@
 git submodule init && git submodule update
-mv ~/.vimrc ~/.vimrc.backup
+if [ -f ~/.vimrc ]; then
+    mv ~/.vimrc ~/.vimrc.backup
+fi
 echo "so ~/.vim/vimrc" > ~/.vimrc
